code refactor. Remove code duplication
