+++ app/__init__.py
@@ -4,7 +4,7 @@ from instance.config import app_config
 from flask_sqlalchemy import SQLAlchemy
 from flask_migrate import Migrate
 from flask_cors import CORS
-from app.v1 import version1
+from app.v1.v1 import version1
 
 
 db = SQLAlchemy()
@@ -17,8 +17,13 @@ def create_app(config_name):
     CORS(app) #makes cross-origin AJAX possible.
     app.config.from_pyfile('config.py')
     app.register_blueprint(version1, url_prefix="/api/v1/")
-    from app.v2 import version2
+    # import all bluprints and register.
+    from app.v2.business import version2
+    from app.v2.review import review
+    from app.v2.auth import auth
     app.register_blueprint(version2, url_prefix="/api/v2/")
+    app.register_blueprint(auth, url_prefix="/api/v2/")
+    app.register_blueprint(review, url_prefix="/api/v2/")
     # initialize extensions
     migrate = Migrate(app, db)
     db.init_app(app)
